# proto/contract.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/contract.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "SegmentsService"
    },
    {
      "name": "UsersService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/example/addUserToSegment": {
      "post": {
        "operationId": "UsersService_AddUserToSegment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddUserToSegmentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddUserToSegmentRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/v1/example/createSegment": {
      "post": {
        "operationId": "SegmentsService_CreateSegment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateSegmentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateSegmentRequest"
            }
          }
        ],
        "tags": [
          "SegmentsService"
        ]
      }
    },
    "/v1/example/deleteSegment/{slug}": {
      "delete": {
        "operationId": "SegmentsService_DeleteSegment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteSegmentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "slug",
            "description": "Slug is a unique segment name that can be considered as an ID.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "SegmentsService"
        ]
      }
    },
    "/v1/example/getActiveSegments/{userId}": {
      "get": {
        "operationId": "UsersService_GetActiveSegments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetActiveSegmentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddUserToSegmentRequest": {
      "type": "object",
      "properties": {
        "slugsAdd": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "slugsDel": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "v1AddUserToSegmentResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1CreateSegmentRequest": {
      "type": "object",
      "properties": {
        "slug": {
          "type": "string",
          "description": "Slug is a unique segment name that can be considered as an ID."
        }
      }
    },
    "v1CreateSegmentResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1DeleteSegmentResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1GetActiveSegmentsResponse": {
      "type": "object",
      "properties": {
        "slugs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}